BMW_ML_Project/requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.3.2
PyYAML==6.0.1

# Serving / API
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
slowapi==0.1.9
python-multipart==0.0.9

# Tracking / UI
mlflow==2.14.1
streamlit==1.35.0

# Validation / logging
structlog==24.1.0

# Dev / tests
pytest==8.2.0
httpx==0.27.0
black==24.4.2
ruff==0.4.2
mypy==1.10.0
